requirements.txt
python-dotenv
chainlit
langchain
langchain-community
langchain-openai
httpx
beautifulsoup4
faiss-cpu


